requirements.txt
# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# Word Document Processing
python-docx==1.1.0

# AI/NLP - Using OpenAI instead of Anthropic
openai==1.12.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1

